Bug #1067
git-svn-id: svn+ssh://svn.creatis.insa-lyon.fr/svn/vip/Portal/trunk@2355 97043df2-9288-40f0-b43d-ea39296b91c7

+++ vip-datamanagement/src/main/java/fr/insalyon/creatis/vip/datamanager/client/view/browser/DataRecord.java
@@ -49,7 +49,7 @@ public class DataRecord extends ListGridRecord {
         this(type, name, 0, "", "", "");
     }
 
-    public DataRecord(String type, String name, int length, String date,
+    public DataRecord(String type, String name, long length, String date,
             String replicas, String permissions) {
         if (name.equals("Trash")) {
             setAttribute("icon", "datamanager/icon-trash");
